refactor: Further refactors for more complex components
