+++ src/components/Atoms/Properties/Templates/TextTemplate.js
@@ -1,6 +1,6 @@
 import React from "react";
 import {useDispatch, useSelector} from "react-redux";
-import {selectProperty, setProperty,} from "../../features/State Management/questionSlice";
+import {selectProperty, setProperty,} from "../../../../features/State Management/questionSlice";
 import {FilledInput, FormControl, InputLabel} from "@material-ui/core";
 
 export const TextProperty = ({
