+++ src/components/Atoms/TextArrayTemplate/EachOptionHides.js
