+++ src/components/Atoms/TextArrayTemplate/EachOptionShows.js
